agent/requirements.txt
streamlit>=1.32.0
requests>=2.31.0
Pillow>=10.2.0
python-dateutil>=2.9.0
tqdm>=4.66.0